+++ src/commands/inputBoxCommand.ts
@@ -1,10 +1,10 @@
 import * as vscode from 'vscode';
-import { ReplaceParameters } from './replaceParameters';
+import { CommandParameters } from './commandParameters';
 
 export class InputBoxCommand implements vscode.Command {
   public static readonly commandName = 'l10nization.inputBox';
 
-  constructor(args: ReplaceParameters[]) {
+  constructor(args: CommandParameters[]) {
     this.title = 'Extract to arb files';
     this.command = InputBoxCommand.commandName;
     this.arguments = args;
@@ -16,5 +16,5 @@ export class InputBoxCommand implements vscode.Command {
 
   tooltip?: string;
 
-  arguments?: ReplaceParameters[];
+  arguments?: CommandParameters[];
 }