+++ src/extension/getConfiguration.ts
@@ -1,0 +1,7 @@
+import * as vscode from 'vscode';
+
+export function getConfiguration(
+  section: string
+): vscode.WorkspaceConfiguration {
+  return vscode.workspace.getConfiguration(section);
+}