+++ src/test/suite/parser.spec.ts
@@ -1,5 +1,8 @@
+import {
+  getStringWithoutEscapes,
+  getVariablesInInterpolation
+} from '../../shared/parser/parser';
 import { expect } from 'chai';
-import { getStringWithoutEscapes } from '../../extension/parser/parser';
 
 function charnL(value: string): string {
   if (value === '') {
@@ -42,6 +45,26 @@ describe('getStringWithoutEscapes', () => {
   tests.forEach(({ actual, expected }) => {
     it(`should return ${charnL(expected)} when ${charnL(actual)}`, () => {
       expect(getStringWithoutEscapes(actual)).to.be.eq(expected);
+    });
+  });
+});
+
+describe('getVariableInInterpolation', () => {
+  const tests = [
+    { actual: 'a $b c', expected: ['b'] },
+    { actual: 'a $b $c d', expected: ['b', 'c'] },
+    { actual: 'a $name d', expected: ['name'] },
+    { actual: 'a $otherName1 d', expected: ['otherName1'] },
+    // eslint-disable-next-line no-template-curly-in-string
+    { actual: 'a $name ${other.name} d', expected: ['name', 'other.name'] },
+    // eslint-disable-next-line no-template-curly-in-string
+    { actual: 'a ${other.toString()} d', expected: ['other.toString()'] }
+  ];
+  tests.forEach(({ actual, expected }) => {
+    it(`should return ${charnL(expected.toString())} when ${charnL(
+      actual
+    )}`, () => {
+      expect(getVariablesInInterpolation(actual)).to.be.all.members(expected);
     });
   });
 });